requirements.txt
pypdf2
langchain
python-dotenv
